+++ scripts/test/SysConfig/OverriddenFileNameGet.t
@@ -15,8 +15,6 @@ use Kernel::System::VariableCheck qw( IsHashRefWithData );
 
 use vars (qw($Self));
 
-use Kernel::Config;
-
 $Kernel::OM->ObjectParamAdd(
     'Kernel::System::UnitTest::Helper' => {
         RestoreDatabase => 1,
@@ -25,6 +23,7 @@ $Kernel::OM->ObjectParamAdd(
 
 my $HelperObject    = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
 my $SysConfigObject = $Kernel::OM->Get('Kernel::System::SysConfig');
+my $ConfigObject    = $Kernel::OM->Get('Kernel::Config');
 
 $HelperObject->ConfigSettingChange(
     Valid => 1,
@@ -38,6 +37,15 @@ $HelperObject->ConfigSettingChange(
         '7' => 2,
         '3' => 1,
     },
+);
+
+my $ArchiveRestoreEffectiveValue = $ConfigObject->Get(
+    'Ticket::EventModulePost'
+)->{'2300-ArchiveRestore'};
+
+$Self->True(
+    $ArchiveRestoreEffectiveValue,
+    'Setting Ticket::EventModulePost###2300-ArchiveRestore is active and has value set.',
 );
 
 my $ExpectedResultRegex = '^Kernel/Config/Files/ZZZZUnitTest\d+\.pm$';
@@ -77,6 +85,64 @@ my @Tests = (
                 '3' => 0,
             },
             UserID => 1,
+        },
+        ExpectedResultRegex => $ExpectedResultRegex,
+    },
+    {
+        Name   => 'Setting Ticket::EventModulePost###2300-ArchiveRestore is not changed yet',
+        Params => {
+            SettingName    => 'Ticket::EventModulePost###2300-ArchiveRestore',
+            EffectiveValue => $ArchiveRestoreEffectiveValue,
+            UserID         => 1,
+        },
+
+        # ExpectedResultRegex => $ExpectedResultRegex,
+    },
+);
+
+for my $Test (@Tests) {
+    my $Result = $SysConfigObject->OverriddenFileNameGet(
+        %{ $Test->{Params} },
+    );
+
+    if ( $Test->{ExpectedResultRegex} ) {
+
+        # We can't compare real file name, since HelperObject uses random numbers to generate it.
+        $Self->True(
+            $Result =~ m{$Test->{ExpectedResultRegex}} // '',
+            "OverriddenFileNameGet() - $Test->{Name} - Check expected result($Test->{ExpectedResultRegex})."
+        );
+    }
+    else {
+        $Self->False(
+            $Result // '',
+            "OverriddenFileNameGet() - $Test->{Name} - not found."
+        );
+    }
+}
+
+# Simulates delete statement.
+$HelperObject->ConfigSettingChange(
+    Valid => 0,
+    Key   => 'Ticket::EventModulePost###2300-ArchiveRestore',
+);
+
+my $ArchiveRestoreEffectiveValue2 = $ConfigObject->Get(
+    'Ticket::EventModulePost'
+)->{'2300-ArchiveRestore'};
+
+$Self->False(
+    $ArchiveRestoreEffectiveValue2,
+    'Setting Ticket::EventModulePost###2300-ArchiveRestore is not active.',
+);
+
+@Tests = (
+    {
+        Name   => 'Setting Ticket::EventModulePost###2300-ArchiveRestore is changed',
+        Params => {
+            SettingName    => 'Ticket::EventModulePost###2300-ArchiveRestore',
+            EffectiveValue => $ArchiveRestoreEffectiveValue,
+            UserID         => 1,
         },
         ExpectedResultRegex => $ExpectedResultRegex,
     },